Strip URIPrefix always
Previously we would do so only when serving static files. Doing it
globally simplifies things. This also fixes an issue where static
requests would fail if our path was not including the prefix.

Also add another config option for a path to the template directory.
Rather than sharing the one for static files served directly with HTTP.
I think this is clearer.