Use GUID in deciding whether to record an item
I check now for whether an item has a GUID. If it does, rely on that
rather than heuristics with publication date.

I encountered a feed with an item that was missed due to the cut off
date logic. While not all feeds have GUIDs, this one did, so if this
GUID logic was in place we'd include it.

As this meant that publication dates get ignored when a feed uses GUIDs,
I've reworked what happens on a feed's first poll. I don't want to mass
import unread items in that case. So now we look for whether it is the
feed's first poll, and if so, record all of the items but set them read.
Subsequent polls will be as usual.